Add new npm-naming exemptions (DefinitelyTyped/DefinitelyTyped#36969)
* Add new npm-naming exemptions

These packages fail the upcoming requirement that the types package
version match a version that exists for the original package.

In other words, these packages have a version, like 0.0 or 1.0, that
their original package doesn't have. The "npm-naming" lint rule will
soon prevent this, so these packages need to be exempt from this rule.

* Restore some useful comments/formatting

* Update required TS version for stale packages